refactor: 重构 Controller 以适配多平台
