+++ electron/main/tasks/autoMessage.ts
@@ -5,9 +5,9 @@ import { IPC_CHANNELS } from 'shared/ipcChannels'
 import { createLogger } from '#/logger'
 import type { Account } from '#/taskManager'
 import { pageManager } from '#/taskManager'
+import { LiveController } from '#/tasks/controller/LiveController'
 import { randomInt, takeScreenshot, typedIpcMainHandle } from '#/utils'
 import windowManager from '#/windowManager'
-import { LiveController } from './Controller'
 import type { BaseConfig } from './scheduler'
 import { TaskScheduler } from './scheduler'
