+++ electron/main/tasks/autoReply.ts
@@ -5,9 +5,9 @@ import { COMMENT_LIST_WRAPPER } from '#/constants'
 import { createLogger } from '#/logger'
 import type { Account } from '#/taskManager'
 import { pageManager } from '#/taskManager'
+import { LiveController } from '#/tasks/controller/LiveController'
 import { sleep, typedIpcMainHandle } from '#/utils'
 import windowManager from '#/windowManager'
-import { LiveController } from './Controller'
 import { AutoReplyPlus } from './autoReplyPlus'
 
 const TASK_NAME = '监听评论'
